AshDispatch.Resource.Dsl.Counter (AshDispatch v0.1.0)

Data structure representing a counter definition from the DSL.

This is the target of the counter entity in the counters section.

Fields

:name - Unique counter identifier in DSL (atom, e.g., :pending_orders_counter)

:trigger_on - List of action names that trigger this counter

:counter_name - Counter name to broadcast (atom)

:resource - Ash resource module to query for counting (optional, defaults to current resource)

:query_filter - Ash.Query filter expression to count items

:audience - Audience type (:user, :admin, or :system)

:invalidates - Frontend query keys to invalidate (list of strings)

Types

t()

@type t() :: %AshDispatch.Resource.Dsl.Counter{
  __spark_metadata__: term(),
  audience: :user | :admin | :system,
  counter_name: atom(),
  invalidates: [String.t()],
  name: atom(),
  query_filter: any(),
  resource: module() | nil,
  trigger_on: [atom()]
}